Moving the bio ner to example.
