+++ forte/processors/__init__.py
@@ -18,4 +18,3 @@ from forte.processors.ner_predictor import *
 from forte.processors.srl_predictor import *
 from forte.processors.text_generation_processor import *
 from forte.processors.vocabulary_processor import *
-from forte.processors.transformers_processor import *
